[RENAME]  : ✏️ 공통으로 사용되는 comment 기능 관련 파일 패키지 분리
